pyproject.toml
[project]
name = "uisurf-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "haruiz", email = "henryruiz22@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "firebase-admin>=6.7.0",
    "google-adk[a2a]>=1.26.0",
    "google-cloud-aiplatform>=1.111.0",
    "google-genai>=1.30.0",
    "pillow>=10.4.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.10.1",
    "email-validator>=2.2.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.34.0",
    "rich>=14.3.3",
]

[project.scripts]
uisurf-app = "uisurf_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "litellm>=1.82.2",
]
